Smart-attendence-system-main/SAMS-starter/backend/requirements.txt
Flask==2.1.3
flask-cors==3.0.10
numpy==1.24.0
pillow==9.4.0
face_recognition==1.3.0
sqlalchemy==1.4.46
